add all_callbacks
